First blind bash at supporting SDC uploads
Explicitly only supports recently uploaded fiels (so no need to merge
with pre-existing data.

Support in this patch is for uploading such data (as part of the
make_info json output or as per-mediafile .sdc files). There is no support
for the creation of the sdc data itself although the patch defines a fairly
simplistic internal format for describing it.

Todos:
* Support time-value claims (statment or qualifier)
* Support mulitple values for same property claim (statment or qualifier)